Began building a proper class structure for the classes I want to implement. Currently, it contains Entities, Props and Actors.
Entites are just transforms and IDs in the world.
Props are entities with models and textures.
Actors are the interactable characters containing movement info and selection capability.

Alot of the changes are mostly the class renamings.

As well as this, added a new helper function that contains code to move towards a certain point at a set speed for a start to unit movement.